+++ src/utils/os_environment_check.py
@@ -10,9 +10,86 @@ from loguru import logger
 from selenium import webdriver
 from selenium.webdriver.chrome.service import Service
 from run import constants
+import zipfile
+import requests
+from utils.download_driver import AutoDownloadChromeDrive, get_chrome_version_from_executable
 
 
-# 主程序
+@logger.catch
+def auto_download_webdriver():
+    """
+
+    :return:
+    """
+    chrome = AutoDownloadChromeDrive()
+    chrome.start()
+
+
+@logger.catch
+def auto_download_chrome():
+    """
+
+    :return:
+    """
+    dest_path = constants.basic_path + "/chrome.exe"  # 对于Windows系统
+    url = search_chrome_download_link('123.0.6312.58')
+    response = requests.get(url, stream=True)
+    if response.status_code == 200:
+        with open(dest_path, 'wb') as file:
+            for chunk in response.iter_content(chunk_size=8192):
+                file.write(chunk)
+        logger.info(f"File downloaded to {dest_path}")
+        extract_zip(dest_path, dest_path)
+        logger.success("unzip success, please open chrome.exe install!")
+    else:
+        logger.error(f"Failed to download file. Status code: {response.status_code}")
+
+
+@logger.catch
+def extract_zip(zip_path, extract_to):
+    """
+
+    :param zip_path:
+    :param extract_to:
+    :return:
+    """
+
+    with zipfile.ZipFile(zip_path, 'r') as zip_ref:
+        zip_ref.extractall(extract_to)
+    logger.info(f"File extracted to {extract_to}")
+
+
+@logger.catch
+def search_chrome_download_link(version):
+    """
+
+    :param version:
+    :return:
+    """
+
+    search_url = f"https://registry.npmmirror.com/-/binary/chromedriver/"
+
+    response = requests.get(search_url)
+    # get version url
+    version_link = None
+    if response.status_code == 200:
+        search_results = response.json()
+        for item in search_results:
+            if version in item['url'] and "LATEST_RELEASE" not in item['url']:
+                version_link = item['url']
+    if version_link:
+        # get win version download link
+        response_link = requests.get(version_link)
+        if response_link.status_code == 200:
+            link_results = response_link.json()
+            for item_link in link_results:
+                if "win" in item_link['name'] or "win" in item_link['url']:
+                    return item_link['url']
+            return None
+    else:
+        return None
+
+
 @logger.catch
 def detect_installed():
     """
@@ -27,6 +104,13 @@ def detect_installed():
         logger.warning("url: https://googlechromelabs.github.io/chrome-for-testing/#stable select correct version to"
                        " download, finish config download webdriver.exe path to chrome_path from "
                        "./config/config.ini file.")
+        chrome_version = get_chrome_version_from_executable()
+        logger.info(f"current google chrome version: {chrome_version}.")
+        if chrome_version:
+            auto_download_webdriver()
+        else:
+            logger.warning("WARNING! you pc not google chrome, will auto download and unzip, please manual install!")
+            auto_download_chrome()
         return False
     return True
 
@@ -59,7 +143,3 @@ def is_chromedriver_installed():
         logger.error("ChromeDriver is not installed or not configured correctly.")
         logger.error("Error message:", str(e))
         return False
-
-#
-# if __name__ == '__main__':
-#     detect_installed()